requirements.txt
discord.py==2.4.0
aiohttp==3.9.1
Pillow
audioop-lts
